# dev-docs/openapi-mvp.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OSMU MVP API",
    "version": "0.1.0-prototype",
    "description": "Machine-readable contract snapshot for the current OSMU prototype. It covers the REST control plane and the MVP S3-compatible alias surface used by the local release gate."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local prototype backend"
    }
  ],
...
  "security": [
    { "bearerAuth": [] }
  ],
  "paths": {
...
    "/api/admin/users": {
      "get": {
        "tags": ["Users", "Admin"],
        "operationId": "getUsers",
        "summary": "List users visible to the administrator",
        "responses": {
          "200": { "$ref": "#/components/responses/JsonList" }
        }
      },
      "post": {
        "tags": ["Users", "Admin"],
        "operationId": "createUser",
        "summary": "Create a user",
        "requestBody": { "$ref": "#/components/requestBodies/JsonObject" },
        "responses": {
          "200": { "$ref": "#/components/responses/JsonOk" },
          "201": { "$ref": "#/components/responses/JsonOk" }
        }
      }
    },
    "/api/admin/users/{userId}/status": {
      "patch": {
        "tags": ["Users", "Admin"],
        "operationId": "updateUserStatus",
        "summary": "Activate or deactivate a user",
        "parameters": [
          { "$ref": "#/components/parameters/UserId" }
        ],
        "requestBody": { "$ref": "#/components/requestBodies/JsonObject" },
        "responses": {
          "200": { "$ref": "#/components/responses/JsonOk" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "osmuAccessKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-OSMU-Access-Key"
      },
      "awsSigV4": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "AWS Signature Version 4 authorization header or equivalent presigned query parameters."
      }
    },
    "parameters": {
...
      "UserId": {
        "name": "userId",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
...
    },
    "requestBodies": {
      "JsonObject": {
        "required": true,
        "content": {
          "application/json": {
            "schema": { "type": "object", "additionalProperties": true }
          }
        }
      },
...
    },
    "responses": {
      "JsonOk": {
        "description": "JSON response",
        "content": {
          "application/json": {
            "schema": { "type": "object", "additionalProperties": true }
          }
        }
      },
      "JsonList": {
        "description": "JSON list response",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ListResponse" }
          }
        }
      },
...
    },
    "schemas": {
...
      "ListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "type": "object", "additionalProperties": true }
          },
          "nextCursor": { "type": "string", "nullable": true }
        }
      },
...
    }
  }
}